Simulink Objective Function
You can now use any Simulink model with Matlab's ga_solver instead of being restricted to only using built function of Matlab.

In this code 'x_base' is the workspace variable that my Simulink Model is using to get a response.  The values of x_base are set to the x values passed into the objective function by the ga_solver. this way when you call a.get('simout') it gets the response with the newly set values of x_base.